requirements.txt
nicegui>=2.5
fastapi>=0.110
uvicorn>=0.30
pandas>=2.2
openpyxl>=3.1
pypdfium2>=4.30
reportlab>=4.2
pillow>=10.4
qrcode>=7.4
python-multipart>=0.0.9
pikepdf>=9.2
pyttsx3>=2.90
python-dotenv>=1.0
requests>=2.31
pyperclip>=1.8
